Add "login" and "authentication"
Allow us to have private content on the site, which is protected behind
a login procedure, of sorts...

This commit also adds a new page to the site: Stasis. This page lists
all of the tracks-in-progress, and gives us a place to consolidate
information about these tracks that we need all the time (such as key,
tempo, and a link to mix).

This information isn't really for anyone outside of the band. To
dissuade anyone from coming across this page, I added some JavaScript
that approximates "security".